[MS ABI] Prevent some expressions from evaluating to a constant
Summary:
In the Microsoft ABI, some expressions containing references to variables
cannot be evaluated as a constant.  These are expressions containing a
conditional, logical and/or, or comma operator with an operand that is a
variable name.

This is observable at the ABI level by whether the compiler would emit a
static initializer for such expressions where normally it would emit a
readonly constant data.  See PR26210 for more details.

Reviewers: rnk

Subscribers: cfe-commits

Differential Revision: http://reviews.llvm.org/D16465